feat(component): Pass element as argument to renderCallback() for easier destructuring.
fixes #958
